requirements.txt
GitPython>=3.1.0
matplotlib>=3.3.0
pytest>=6.0.0
